Move actions to new folder, restructure inheritance
