+++ robots/ranseur/src/main/cpp/ranseurcameratracker/SetThresholdAction.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "RanseurCameraTracker.h"
-#include <Action.h>
+#include <actions/Action.h>
 
 namespace xero {
     namespace ranseur {
