+++ robots/ranseur/src/main/cpp/ranseuroi/RanseurOISubsystem.cpp
@@ -5,7 +5,7 @@
 #include <Robot.h>
 #include <SettingsParser.h>
 #include <oi/DriverGamepad.h>
-#include <SequenceAction.h>
+#include <actions/SequenceAction.h>
 #include <frc/DriverStation.h>
 #include <iostream>
 #include <cmath>
