+++ xerolibs/xerobase/DetectAutoSequence.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Action.h"
+#include "actions/Action.h"
 #include <memory>
 
 namespace xero
