+++ xerolibs/xerobase/Subsystem.cpp
@@ -1,5 +1,5 @@
 #include "Subsystem.h"
-#include "Action.h"
+#include "actions/Action.h"
 #include "Robot.h"
 #include "basegroups.h"
 #include <MessageLogger.h>
