+++ xerolibs/xerobase/Subsystem.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Action.h"
+#include "actions/Action.h"
 #include "LoopType.h"
 #include <memory>
 #include <map>
