+++ xerolibs/xerobase/actions/DispatchAction.cpp
@@ -1,4 +1,4 @@
-#include "DispatchAction.h"
+#include "actions/DispatchAction.h"
 #include "basegroups.h"
 #include "Subsystem.h"
 #include "Robot.h"
