+++ xerolibs/xerobase/actions/DispatchAction.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Action.h"
+#include "actions/Action.h"
 #include "Subsystem.h"
 
 /// \file
@@ -10,7 +10,7 @@ namespace xero {
         /// \brief this is a an action that wraps a subsystem and an action to assign to a subsystem
         class DispatchAction : public Action {
         public:
-            /// \brief create a new wrapper to wrap a subsystem and an assocaited action
+            /// \brief create a new wrapper to wrap a subsystem and an associated action
             /// \param subsystem the subsystem to assign an action to
             /// \param action the action to assign to the subsystem
             /// \param block if true, wait for the subsystem to complete the action before returning isDone() true
@@ -36,7 +36,12 @@ namespace xero {
             /// \brief returns true if the subsystem did not accept the action
             /// \returns true if the action was not accepted by the subsystem
             bool isDenied() { return denied_ ;}
-            
+
+            /// \returns A pointer to the subsystem.
+            SubsystemPtr getSubsystem() { return subsystem_; }
+
+            /// \returns A pointer to the action.
+            ActionPtr getAction() { return action_; }
         private:
             //
             // The subsystem to assign the action to
