+++ xerolibs/xerobase/actions/ParallelAction.cpp
@@ -1,4 +1,4 @@
-#include "ParallelAction.h"
+#include "actions/ParallelAction.h"
 #include <iostream>
 
 namespace xero {
