+++ xerolibs/xerobase/actions/TerminateAction.cpp
